frontend.py
This code defines a simple screen with a card containing a text field and a button for the user to enter a video URL and scan a car plate. The scan_plate function is called when the user clicks the button, but it still needs to be implemented to actually scan the plate. You can replace the TODO comment with your own implementation.